+++ rt/exceptions.py
@@ -1,7 +1,5 @@
 """Exceptions collection for the rt library."""
 
-import typing
-
 
 class RtError(Exception):
     """Super class of all Rt Errors."""
@@ -18,7 +16,7 @@ class NotAllowedError(RtError):
 class UnexpectedResponseError(RtError):
     """Exception raised when unexpected HTTP code is received."""
 
-    def __init__(self, message: str, status_code: typing.Optional[int] = None, response_message: typing.Optional[str] = None) -> None:
+    def __init__(self, message: str, status_code: int | None = None, response_message: str | None = None) -> None:
         """Initialization."""
         super().__init__(message)
         self.status_code = status_code